requirements.txt
# Core dependencies
python-docx==1.1.0
PyPDF2==3.0.1
pandas==2.1.4
openpyxl==3.1.2

# NLP and ML
spacy==3.7.2
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.9.1
faiss-cpu==1.13.2

# OpenAI (for GPT extraction)
openai>=1.10.0

# UI
customtkinter>=5.2.0
pynput>=1.7.6

# Utilities
python-dotenv==1.0.0
numpy==1.26.3
tqdm==4.66.1
watchdog>=3.0.0
